% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnabin_functions.R
\name{align_mafft}
\alias{align_mafft}
\title{Align with \code{add} from MAFFT}
\usage{
align_mafft(fas1 = "data/MSM_CRF02AG_splits/ref_MSM_CRF02AG.fas",
  fas2 = "data/MSM_CRF02AG_splits/uk_MSM_CRF02AG.fas",
  out = "data/MSM_CRF02AG_splits/MSM_CRF02AG_nodrm_ref.fas", verbose = TRUE)
}
\arguments{
\item{fas1}{Path to MSA in fasta format (small)}

\item{fas2}{Path to MSA in fasta format (big)}

\item{out}{Path to output in fasta format}

\item{verbose}{explain stuff}
}
\value{
Path to output fasta file
}
\description{
Add \code{fas1} to \code{fas2}. Need \code{mafft/7.271}
}
