% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnabin_functions.R
\name{bigphylo_drm}
\alias{bigphylo_drm}
\title{Apply \code{big.phylo::seq.rm.drugresistance}}
\usage{
bigphylo_drm(infile = "test.rds", wait = 0, verbose = TRUE)
}
\arguments{
\item{infile}{Path of RDS file containing DNAbin object}

\item{wait}{sleep time option to allow qsub dependency}

\item{verbose}{gives computation time}
}
\value{
Path of RDS file containing sequences without DRM
}
\description{
Convert in place \code{*.rds} to \code{nodrm_*.rds}
}
