% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnabin_functions.R
\name{blast_dnabin}
\alias{blast_dnabin}
\title{BLAST sequences against database}
\usage{
blast_dnabin(seq, path.database, metric = "bitscore", target_seqs = 5,
  subtype = FALSE, verbose = TRUE)
}
\arguments{
\item{seq}{Input seqs in dnabin format}

\item{path.database}{Path to database in fasta}

\item{metric}{criterium for best hit. Among \code{c('pident', 'bitscore', 'evalue')} [default = \code{bitscore}]}

\item{target_seqs}{Maximum number of aligned sequences to keep, per query [default = 5]}

\item{subtype}{Print distribution of subtypes of hits}

\item{verbose}{explain stuff}
}
\value{
Output a list of tables
}
\description{
Runs \code{blastn} of \code{seq}(query sequences) against a database. Returns tables of
\describe{
\item{all_hits}{all matches per query sequence}
\item{best_hits}{best match per query sequence}
\item{unique_hits}{unique database sequences found as best match}
}
}
