% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnabin_functions.R
\name{blast_splits}
\alias{blast_splits}
\title{BLAST splits against database}
\usage{
blast_splits(infile = "test.rds",
  database = "blastDB/hiv1_FLT_noUK_nodrm.fasta", target_seqs = 5,
  verbose = TRUE)
}
\arguments{
\item{infile}{Path to query sequence (split) in RDS format}

\item{database}{Path to database in fasta}

\item{target_seqs}{Maximum number of aligned sequences to keep, per query [default = 5]}

\item{verbose}{explain stuff}
}
\value{
Path to output list of tables
}
\description{
Remove hxb2, apply blastn, output best hits
}
