% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnabin_functions.R
\name{check.seqs.input}
\alias{check.seqs.input}
\title{Check if input is sequences (DNAbin in compressed R file)}
\usage{
check.seqs.input(path.seqs = "msa.rds")
}
\arguments{
\item{path.seqs}{Path to R file}
}
\value{
DNAbin object
}
\description{
Check if input is \code{rds} or \code{rda}, and contains a DNAbin object
}
