% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_st}
\alias{check_st}
\title{Check subtype from LANL sequences labels}
\usage{
check_st(x, k = 6)
}
\arguments{
\item{x}{DNAbin object or table of BLAST results}

\item{k}{number of subtype to return}
}
\value{
proportion of \code{k} most frequent
}
\description{
Check subtype from LANL sequences labels
}
