% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{count_bstrees}
\alias{count_bstrees}
\title{Count files existing}
\usage{
count_bstrees(path = "data/BSTREES",
  pattern = "ExaML_result.*finaltree\\\\.\\\\d{3}$", n_expected = 1:100)
}
\arguments{
\item{path}{Path to trees}

\item{pattern}{For list.files regex}

\item{n_expected}{vector of numbers of bootstrap expected}
}
\value{
list of table(group, n, file), number of files by group, number missing
}
\description{
Count files existing
}
