% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnabin_functions.R
\name{dashToN}
\alias{dashToN}
\title{Substitute characters for gaps in DNAbin object}
\usage{
dashToN(s, a = "-", b = "n")
}
\arguments{
\item{s}{Object in DNAbin format}

\item{a}{old gap character (default = '-')}

\item{b}{new gap character (default = 'N')}
}
\value{
DNAbin object
}
\description{
By default substitute dashes to 'N' either for list or matrix forms of DNAbin
}
