% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnabin_functions.R
\name{dates_of_rtree}
\alias{dates_of_rtree}
\title{Find dates of tree for UK and LANL sequences}
\usage{
dates_of_rtree(TREE, ALLDATES_LANL = get_lanl_dates(),
  ALLDATES_UK = get_uk_dates(), rooting = FALSE, verbose = TRUE)
}
\arguments{
\item{TREE}{Path to tree}

\item{ALLDATES_LANL}{Path to dates file from LANL sequences}

\item{ALLDATES_UK}{Path to dates from UK sequences}

\item{rooting}{rooting on outgroup}

\item{verbose}{Details}
}
\value{
List of (tree, dates, name of tree, number tips)
}
\description{
Find dates of tree for UK and LANL sequences
}
