% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnabin_functions.R
\name{dnabinToMatrix}
\alias{dnabinToMatrix}
\title{Convert DNAbin to matrix form}
\usage{
dnabinToMatrix(l, gaps = "-")
}
\arguments{
\item{l}{Object in DNAbin format}

\item{gaps}{format of gaps in MSA (default = '-')}
}
\value{
DNAbin in matrix format
}
\description{
Padding with gaps to match max(length)
}
