% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{do_examl}
\alias{do_examl}
\title{Compute ExaML tree}
\usage{
do_examl(FAS = "data/MSM_CRF02AG_splits/MSM_CRF02AG_nodrm_ref_og.fas",
  DIR = "data/ML_TREE0", bsdebut = 0, bsend = 2, mem = "32GB", q = NA)
}
\arguments{
\item{FAS}{Path to fasta MSA}

\item{DIR}{Folder for input MSA and output tree}

\item{bsdebut}{Boostrap debut}

\item{bsend}{Bootstrap end}

\item{mem}{memory}

\item{q}{HPC queue}
}
\value{
If on HPC, output PBS_JOBID number(s)
}
\description{
Either in HPC or locally
}
