% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{do_lsd}
\alias{do_lsd}
\title{LSD on HPC}
\usage{

  do_lsd(TREE = "data/ML_TREE0/ExaML_result.MSM_CRF02AG_nodrm_ref_G.finaltree.000",
  outgroup = TRUE, ...)
}
\arguments{
\item{TREE}{Path to tree}

\item{outgroup}{Outgroup option of LSD (LSD outputs a tree without outgroup)}

\item{...}{Passed to hpc.cmd.funr (i.e. hpc.q = 'pqeelab', verbose = FALSE)}
}
\description{
LSD on HPC
}
