% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{do_rtt}
\alias{do_rtt}
\title{RTT filtering}
\usage{

  do_rtt(TREE = "data/ML_TREE0/ExaML_result.MSM_CRF02AG_nodrm_ref_G.finaltree.000",
  ...)
}
\arguments{
\item{TREE}{Path to tree}

\item{...}{Passed to hpc.cmd.funr (i.e. hpc.q = 'pqeelab', verbose = FALSE)}
}
\description{
RTT filtering
}
