% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{do_tmrca_outliers}
\alias{do_tmrca_outliers}
\title{Finding tmrca outliers on HPC}
\usage{
do_tmrca_outliers(PATHMSA = "data/ML_TREE0/MSM_CRF02AG_nodrm_ref_og.R", ...)
}
\arguments{
\item{PATHMSA}{Path of msa file}

\item{...}{Passed to hpc.cmd.funr (i.e. hpc.q = 'pqeelab', verbose = FALSE)}
}
\description{
See find_tmrca_outliers
}
