% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{do_tmrca_outliers_splits}
\alias{do_tmrca_outliers_splits}
\title{Finding tmrca outliers on HPC with splitting step}
\usage{
do_tmrca_outliers_splits(pathmsa, ...)
}
\arguments{
\item{pathmsa}{Path of msa file
e.g. "data/MSM_CRF02AG_splits/MSM_CRF02AG_nodrm_ref_og.fas"}

\item{...}{Passed to hpc.cmd.funr (i.e. hpc.q = 'pqeelab', verbose = FALSE)}
}
\description{
See find_tmrca_outliers
}
