% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assortativity.R
\name{edgelist_to_mixingmatrix}
\alias{edgelist_to_mixingmatrix}
\title{Weighted edgelist to mixing matrix}
\usage{
edgelist_to_mixingmatrix(df, transpose = TRUE)
}
\arguments{
\item{df}{dataframe in form [from, to, weight]}

\item{transpose}{If TRUE, \code{from} in column and \code{to} in row}
}
\description{
Add weights for each combination of \code{from} and \code{to} pairs in a normalised matrix. Used for mixing matrix of probability that there is one transmission between donor_j (column) to recipient_i (row) ( total sums to 1 ). Or donor = row and recip = column if \code{transpose = FALSE}.
}
\examples{
( df <- data.frame(from = c('a','a','b','b'), to = c('a','b','b','b') , ip = c(.2,.8,.5,.5)) )
( m <- edgelist_to_mixingmatrix(df) ) # mixing matrix
m / rowSums(m)[row(m)] # If transpose = TRUE,
# P(donor_j|recipient_i) or
# (probability that recipient row has donor column).
# Every row sums to 1
}
