% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnabin_functions.R
\name{filter_tree_outliers}
\alias{filter_tree_outliers}
\title{Remove RTT and MRCA outliers from tree}
\usage{
filter_tree_outliers(TREE, outgroup = TRUE, OUTFILE = NULL,
  rooting = TRUE, verbose = TRUE)
}
\arguments{
\item{TREE}{Path to tree}

\item{outgroup}{Logical indicating that rooting is on outgroup of sequences}

\item{OUTFILE}{Path to save list of outliers, in rds}

\item{rooting}{passed to dates_of_rtree}

\item{verbose}{stuff done}
}
\value{
List of (group name, path to filtered tree, ntips original tree, ntips filtered tree, n rtt outliers, n mrca outliers, proportion tips filtered)
}
\description{
Remove RTT and MRCA outliers from tree
}
