% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnabin_functions.R
\name{filter_tree_outliers2}
\alias{filter_tree_outliers2}
\title{DO NOT USE: Remove RTT and MRCA outliers from tree + branch length outliers}
\usage{
filter_tree_outliers2(TREE, outgroup = TRUE, OUTFILE = NULL,
  rooting = TRUE, verbose = TRUE)
}
\arguments{
\item{TREE}{Path to tree}

\item{outgroup}{Logical indicating that rooting is on outgroup of sequences}

\item{OUTFILE}{Path to save list of outliers, in rds}

\item{rooting}{passed to dates_of_rtree}

\item{verbose}{stuff done}
}
\description{
DO NOT USE: Remove RTT and MRCA outliers from tree + branch length outliers
}
