% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnabin_functions.R
\name{find.clades.0}
\alias{find.clades.0}
\title{Find clades of state = 0 descending from state = 1}
\usage{
find.clades.0(tree, tip.state, verbose = TRUE, plots = FALSE)
}
\arguments{
\item{tree}{phylo tree}

\item{tip.state}{0/1 for UK/non-UK}

\item{verbose}{details}

\item{plots}{plot colored tree}
}
\description{
Find clades of state = 0 descending from state = 1
}
