% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnabin_functions.R
\name{find.rtt.outliers2}
\alias{find.rtt.outliers2}
\title{C++ version of Root-to-tip regression outliers}
\usage{
find.rtt.outliers2(tree, dates, mult = 2, title = NULL, verbose = TRUE,
  plots = FALSE)
}
\arguments{
\item{tree}{phylo tree}

\item{dates}{with tip names}

\item{mult}{multiplier of sd to define outliers}

\item{title}{title of plot}

\item{verbose}{explain}

\item{plots}{plot}
}
\description{
with dist2root.r
}
