% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnabin_functions.R
\name{find_clades_slice}
\alias{find_clades_slice}
\title{Extract clades slicing from root}
\usage{
find_clades_slice(tres, UB = 1000, DSLICE = 1)
}
\arguments{
\item{tres}{Phylo tree}

\item{UB}{Upper bound of clade size}

\item{DSLICE}{Height above the root at which tree is sliced}
}
\description{
Recursively slice big tree to pull out clades within given size bounds
}
