% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnabin_functions.R
\name{find_db}
\alias{find_db}
\title{Find BLAST database}
\usage{
find_db(infile = "path.msa", dbdir = "blastDB", verbose = TRUE)
}
\arguments{
\item{infile}{Path to MSA}

\item{dbdir}{Folder of BLAST databases}

\item{verbose}{explain stuff}
}
\value{
Path to database
}
\description{
Find subtype on infile name and then database
}
