% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnabin_functions.R
\name{find_mrca_outliers}
\alias{find_mrca_outliers}
\title{Find tips that should not be ancestral}
\usage{
find_mrca_outliers(tree)
}
\arguments{
\item{tree}{dated tree}
}
\value{
problematic tip labels
}
\description{
Find tips producing nodes ancestral to MRCA from global external sequences
}
