% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsd_function.R
\name{find_tmrca_outliers}
\alias{find_tmrca_outliers}
\title{Check which seqs generate outliers in tmrca}
\usage{
find_tmrca_outliers(pathmsa = "data/ML_TREE0/MSM_CRF02AG_nodrm_ref_og.R")
}
\arguments{
\item{pathmsa}{msa file in dnabin (*.R) or fasta (*.fas/fasta) format}
}
\description{
Test UK sequences one-by-one with group of lanl seq closest to outgroup
}
