% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnabin_functions.R
\name{get_clades}
\alias{get_clades}
\title{Exclude treated patients and find clades}
\usage{
get_clades(lsd_result, not_naive = NULL, plots = FALSE, verbose = TRUE)
}
\arguments{
\item{lsd_result}{LSD result as list from which [[2]] = dated tree}

\item{not_naive}{vector of non naive test index}

\item{plots}{Plot of colored tree(passed to \code{find.clades.0})}

\item{verbose}{details}
}
\description{
Exclude treated patients and find clades
}
