% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_functions.R
\name{get_ip}
\alias{get_ip}
\title{Compute infector probabilities and bind all clade results from tree}
\usage{
get_ip(pathclade, parms_SA = get_sa_parms_range(), outdir = "data/IPS",
  verbose = TRUE)
}
\arguments{
\item{pathclade}{Path to list of clades}

\item{parms_SA}{Parameters from \code{\link{get_sa_parms_range}}}

\item{outdir}{Create directory in which to save results}

\item{verbose}{print}
}
\value{
List of [(inc, prev), dataframe(donor, recip, ip)]
}
\description{
Compute infector probabilities and bind all clade results from tree
}
