% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_bash.R
\name{get_jobid}
\alias{get_jobid}
\title{Get PBS_JOBID from qsub query}
\usage{
get_jobid(lst = "script", array = TRUE, qstat = FALSE, verbose = TRUE)
}
\arguments{
\item{lst}{List (or list of list) of qsub}

\item{array}{array jobs notation}

\item{qstat}{print qstat}

\item{verbose}{print stuff}
}
\value{
Numeric of JOBID in form: single integer, array notation x:y, or vector if array = FALSE
}
\description{
Check depth of list to find \code{jobid} object
}
