% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_lanl_dates}
\alias{get_lanl_dates}
\title{Extract and format dates from sequences labels}
\usage{
get_lanl_dates(lstinfile = c("db_hiv1_A_noUK", "db_hiv1_B_noUK",
  "db_hiv1_CRF02AG_noUK", "db_hiv1_C_noUK"), outfile = "data/lanl_dates.rds")
}
\arguments{
\item{lstinfile}{List of data files (in package) containing LANL sequences}

\item{outfile}{Path to named vector of dates (numeric in year; for LANL, midpoint year of accession)}
}
\value{
Path to output
}
\description{
For dating trees
}
