% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_functions.R
\name{get_sa_parms_range}
\alias{get_sa_parms_range}
\title{Default parameters for source attribution in MSM.}
\usage{
get_sa_parms_range(filename = NA)
}
\arguments{
\item{filename}{Whether and where to save file}
}
\description{
Surveillance data taken from Brown, 2017 and Yin, 2014.
\itemize{
\item For incidence in MSM:
"Following this adjustment, the estimated number of infections acquired per year rose from around 2,200 infections (95\% credible interval (CrI)1,800 to 2,500) in 2007 to a peak of 2,800 (CrI 2,300 to 3,200) in 2012 before falling to 1,700 (CrI 900 to 2,700) in 2016"
\item Let's assume in the period preceding last sequence (2013) an incidence of 2800 [2300 - 3200].
\item For prevalence, from Yin et al. 2014, MSM living with HIV in 2013 = 43,500 [40,200 - 48,200]
\item Let's further assume that incidence and prevalence are normally distributed and credible interval bounds are equal to 95\% confidence interval bounds, so that \code{sd = (up - lo)/(2 * qnorm(.975))}
\item Hence, values of incidence / prevalence can be drawn from normal distributions
}
}
\references{
1. Brown AE, Kirwan P, Chau C, Khawam J, Gill ON, Delpech VC. Towards elimination of HIV transmission AIDS and HIV related deaths in the UK - 2017 report. Public Health England; 2017.
\url{https://www.gov.uk/government/uploads/system/uploads/attachment_data/file/662306/Towards_elimination_of_HIV_transmission_AIDS_and_HIV_related_deaths_in_the_UK.pdf}
2. Yin Z, Brown AE, Hughes G, Nardone A, Gill ON, Delpech VC, et al. HIV  in  the  United  Kingdom  2014  Report:  data  to  end  2013. London: Public Health England; 2014.
\url{https://www.gov.uk/government/uploads/system/uploads/attachment_data/file/401662/2014_PHE_HIV_annual_report_draft_Final_07-01-2015.pdf}
}
