% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_uk_dates}
\alias{get_uk_dates}
\title{Extract and format dates from sequences labels}
\usage{
get_uk_dates(outfile = "data/uk_dates.rds")
}
\arguments{
\item{outfile}{Path to named vector of dates (numeric in year since 01-01-000)}
}
\value{
Path to output
}
\description{
For dating trees
}
