% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnabin_functions.R
\name{grubbs.flag}
\alias{grubbs.flag}
\title{Iterative grubbs outliers searching}
\usage{
grubbs.flag(x)
}
\arguments{
\item{x}{named vector of values}
}
\value{
data.frame of value with Outlier logical
}
\description{
Iterative grubbs outliers searching
}
