% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assortativity.R
\name{interpret_mat_gg}
\alias{interpret_mat_gg}
\title{Series of 4 heatmaps and assortativity coefficient.}
\usage{
interpret_mat_gg(l_all_mixmat, var, var_order = NULL, group, ...)
}
\arguments{
\item{l_all_mixmat}{List of mixing matrices computed by \code{\link{edgelist_to_mixingmatrix}}, by variable and by group}

\item{var}{First list level of \code{l_all_mixmat}}

\item{var_order}{Optional named list (by variable) of row/col names to reorder matrices}

\item{group}{Second list level of \code{l_all_mixmat}}

\item{...}{option to \code{\link{level_gg}}, i.e. contrast = 2, transpose = FALSE, tl = "something", abbrev_y = abbreviate}
}
\description{
\itemize{
\item mixing: is the probability that there is one transmission between donor of type j to a recipient of type i (cumulative infector probabilities, sums to 1).
\item std_recip: is the probability for a recipient of type i to have a donor of type j (Each row sums to 1).
\item std_donor: is the probability for a donor of type j to transmit to a recipient of type i (Each col sums to 1).
\item assortment: is the excess assortment between types relative to random allocation. Summarized by Newman's assortativity coefficient: r.
}
}
