% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_ip.R
\name{ip_by_var}
\alias{ip_by_var}
\title{Edge list of infector probabilities by patients variable}
\usage{
ip_by_var(iplist, df, var)
}
\arguments{
\item{iplist}{list of ip by patient}

\item{df}{patient data}

\item{var}{variable}
}
\description{
Edge list of infector probabilities by patients variable
}
