% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assortativity.R
\name{level_gg}
\alias{level_gg}
\title{Heatmap from matrix with ggplot}
\usage{
level_gg(mat, contrast = 1, title = NULL, legend = FALSE,
  transpose = TRUE, labx = "donor", laby = "recipient", labz = "ip",
  rotate_y = TRUE, abbrev = identity, values = FALSE, fmt = identity,
  textsize = NULL, col1 = "red", col2 = "yellow")
}
\arguments{
\item{mat}{matrix}

\item{contrast}{change heatmap contrast (independent of plotted values)}

\item{title}{main title}

\item{legend}{logical for legend (if contrast transformation, legend shows untransformed values)}

\item{transpose}{transpose matrix}

\item{labx}{label x}

\item{laby}{label y}

\item{labz}{label of values, passed to legend}

\item{rotate_y}{rotate y-axis labels}

\item{abbrev}{use abbreviate or identity}

\item{values}{write values (untransformed by contrast) in cells}

\item{fmt}{format values, i.e. round, signif, etc. [default: identity]}

\item{textsize}{size of values}

\item{col1}{color low values [default: red]}

\item{col2}{color high values [default: yellow]}
}
\description{
Forced to square rendering
}
\examples{
m <- matrix(1:9, nrow=3, byrow = TRUE)
colnames(m) <- paste0("Column", 1:3)
rownames(m) <- paste0("Row long", 1:3)
m
level_gg(m)
level_gg(m, legend = FALSE, values = TRUE)
}
\seealso{
\code{\link{mat_to_heatmap}}
}
