% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{load_msa}
\alias{load_msa}
\title{Load MSA in various formats}
\usage{
load_msa(pathmsa, verbose = FALSE)
}
\arguments{
\item{pathmsa}{Path to file}

\item{verbose}{print content of seq}
}
\description{
load \code{seq} object from file in .R, .fas/fasta, .rds. Remove HXB2 if any
}
