% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnabin_functions.R
\name{lsd.examl.tree}
\alias{lsd.examl.tree}
\title{Date tree with LSD}
\usage{
lsd.examl.tree(TREE, dates = NA, rooting = TRUE, outgroup = TRUE,
  ci = "-f 10", lsd.version = "0.3", verbose = TRUE)
}
\arguments{
\item{TREE}{Path to tree}

\item{dates}{Dates of tips [default = NA]. By default found by \code{dates_of_tree}}

\item{rooting}{Logical indicating that rooting is done on outgroup of sequences (passed to \code{dates_of_rtree})}

\item{outgroup}{Outgroup option of LSD (LSD outputs a tree without outgroup)}

\item{ci}{option for confidence interval}

\item{lsd.version}{[default = 0.3 (beta)]}

\item{verbose}{Details}
}
\value{
NULL. Save list of [[1]] metrics and [[2]] dated tree
}
\description{
Date tree with LSD
}
