% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnabin_functions.R
\name{lsd.probs}
\alias{lsd.probs}
\title{Find ancient nodes with uk only descendants}
\usage{
lsd.probs(lst = "lsd_results_og_nooout")
}
\arguments{
\item{lst}{List of LSD results}
}
\description{
having dates of each node
}
