% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsd_function.R
\name{lsd_fn}
\alias{lsd_fn}
\title{Generic LSD usage}
\usage{
lsd_fn(tr, dates = NULL, outgroup = NULL, root.constraint = NULL,
  verbose = TRUE, ci = "-f 10")
}
\arguments{
\item{tr}{Phylo tree}

\item{dates}{Dates of tips, by default extracted from accession label [default=NULL]}

\item{outgroup}{Labels of outgroup tips}

\item{root.constraint}{In the form \code{root.constraint = c("mrca(a,b,c,d)", "b(2000,2001)")}}

\item{verbose}{Write details}

\item{ci}{Confidence interval calculation (\code{ci = '-f 10'} or \code{ci = NULL})}
}
\description{
Version 0.3beta
}
