% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnabin_functions.R
\name{lsd_to_clades}
\alias{lsd_to_clades}
\title{Apply \code{\link{find_clades_slice}} on one LSD result file
Remove treated patients sequences and global sequences}
\usage{
lsd_to_clades(lsd_result, max.clade.size = 1000, not_naive = NULL,
  rm.global = TRUE, verbose = TRUE)
}
\arguments{
\item{lsd_result}{LSD result as list from which [[2]] = dated tree}

\item{max.clade.size}{Upper bound of clade size}

\item{not_naive}{vector of non naive test index}

\item{rm.global}{remove global seqs}

\item{verbose}{details}
}
\description{
Apply \code{\link{find_clades_slice}} on one LSD result file
Remove treated patients sequences and global sequences
}
