% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnabin_functions.R
\name{mafft_it}
\alias{mafft_it}
\title{Align new sequences to existing alignment}
\usage{
mafft_it(seq = ape::read.dna(system.file("ext", "hxb2.txt", package = "garel",
  mustWork = TRUE), format = "fasta"), profile,
  args = "--auto --thread -1 --keeplength --add", verbose = TRUE)
}
\arguments{
\item{seq}{new sequences in DNAbin format (default = hxb2)}

\item{profile}{existing alignment (profile in MAFFT jargon) in DNAbin format}

\item{args}{arguments for MAFFT v7. Use \code{--thread 16} in HPC. See \url{https://mafft.cbrc.jp/alignment/server/add.html} and \url{https://mafft.cbrc.jp/alignment/software/addsequences.html}}

\item{verbose}{print stderr}
}
\value{
DNAbin object
}
\description{
With MAFFT v7.
}
