% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnabin_functions.R
\name{make_blast_db}
\alias{make_blast_db}
\title{Make blast DB}
\usage{
make_blast_db(pathfile, pathout = "blastDB")
}
\arguments{
\item{pathfile}{Path of RDS files containing sequences}

\item{pathout}{Folder to store BLAST databases}
}
\description{
Need an installation of ncbi-blast. Turns a RDS file containing MSA into a BLAST database in fasta format
}
