% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assortativity.R
\name{mat2assortmat}
\alias{mat2assortmat}
\title{Assortativity matrix}
\usage{
mat2assortmat(mat)
}
\arguments{
\item{mat}{mixing matrix}
}
\description{
Excess assortment between categories relative to random allocation
}
