% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{mat_to_bar}
\alias{mat_to_bar}
\title{Bar plot from mixing matrix}
\usage{
mat_to_bar(m, title = NULL, x_lab = "recipient", y_lab = "donor",
  z_lab = "ip", horiz = TRUE, values = FALSE, rotate_y = TRUE,
  abbrev = identity)
}
\arguments{
\item{m}{matrix}

\item{title}{title}

\item{x_lab}{rows label}

\item{y_lab}{columns label}

\item{z_lab}{values label}

\item{horiz}{horizontal TRUE (default), vertical FALSE}

\item{values}{show values in or above bar, in black or white [default=FALSE]}

\item{rotate_y}{y-axis labels vertical for vertical plot [default=TRUE]}

\item{abbrev}{optionally pass \code{abbreviate} to axes labels [default=identity]}
}
\description{
Representation of matrices standardized by row / column in horizontal / vertical bar density
}
\examples{
k <- 4
m <- matrix(1:k^2/k^2, nrow = k, byrow = TRUE)
rownames(m) <- paste0("R", 1:k)
colnames(m) <- paste0("C", 1:k)
m
mat_to_bar(m = m, horiz = TRUE, values = TRUE)
mat_to_bar(m = m, horiz = FALSE, values = TRUE)
}
