% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assortativity.R
\name{mat_to_heatmap}
\alias{mat_to_heatmap}
\title{Heatmap from matrix in base R}
\usage{
mat_to_heatmap(m, xlab = "", ylab = "", values = TRUE, ...)
}
\arguments{
\item{m}{matrix}

\item{xlab}{label for columns}

\item{ylab}{label for rows}

\item{values}{display values in cells}

\item{...}{For title, color, etc. plot options}
}
\description{
Heatmap from matrix in base R
}
\examples{
m <- matrix(1:9, nrow=3, byrow = TRUE)
colnames(m) <- paste0("C", 1:3)
rownames(m) <- paste0("R", 1:3)
m
mat_to_heatmap(m)
}
\seealso{
\code{\link{level_gg}}
}
