% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assortativity.R
\name{mat_to_igraph}
\alias{mat_to_igraph}
\title{Bidirectional weighted graph from matrix}
\usage{
mat_to_igraph(m, layout = NULL, labels = FALSE, ...)
}
\arguments{
\item{m}{adjacency matrix}

\item{layout}{default: layout.fruchterman.reingold}

\item{labels}{weights as labels}

\item{...}{passed to plot.igraph.}
}
\description{
With \code{igraph}. See net0_sims_slv.R
}
\examples{
k <- 3
m <- matrix(runif(k*k,0,1), nrow = k, byrow = TRUE)
colnames(m) <- rownames(m) <- LETTERS[1:k]
m
mat_to_igraph(m)
}
