% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_ip.R
\name{mean_ip_per_bs}
\alias{mean_ip_per_bs}
\title{Mean infector probabilities by bootstrap}
\usage{
mean_ip_per_bs(pathip, outfile = NULL)
}
\arguments{
\item{pathip}{List by group of infector probability files}
}
\description{
Mean infector probabilities by bootstrap
}
