% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnabin_functions.R
\name{merge_blasts}
\alias{merge_blasts}
\title{Merge BLAST splits against database}
\usage{
merge_blasts(nblasts = 0, indir = "data/MSM_CRF02AG_splits",
  outfile = "data/MSM_CRF02AG_splits/MSM_CRF02AG_nodrm.rds", verbose = TRUE)
}
\arguments{
\item{nblasts}{Optional number of splits to wait for before merging}

\item{indir}{Folder containing splits of BLAST output}

\item{outfile}{Path to output file based on original MSA name}

\item{verbose}{explain stuff}
}
\value{
Table of unique hits
}
\description{
Merge split BLAST output with best hits
}
