% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnabin_functions.R
\name{merge_out_tmrca}
\alias{merge_out_tmrca}
\title{Parse results form fast_tmrca}
\usage{
merge_out_tmrca(indir = "data/MSM_CRF02AG_splits", outfile = "out_tmrca")
}
\arguments{
\item{indir}{Path to individual tmrca assesments}

\item{outfile}{Name of outfile containing outliers of tmrca}
}
\description{
Used in \code{\link{do_tmrca_outliers_splits}}. Look for Grubbs outliers of TMRCA
}
