% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnabin_functions.R
\name{merge_splits}
\alias{merge_splits}
\title{Merge split sequences}
\usage{
merge_splits(nsplits = 0, indir = "data/MSM_CRF02AG_splits",
  ptrn = ".*split.*_nodrm.rds", outname = "merged_seqs_nodrm",
  verbose = TRUE)
}
\arguments{
\item{nsplits}{Optional number of splits to wait for before merging}

\item{indir}{Path of RDS files containing split sequences}

\item{ptrn}{Regex pattern for selecting which files are splits [default = \code{"nodrm.*\\.rds$"}]}

\item{outname}{Name of output [default = \code{merged_seqs_nodrm}]}

\item{verbose}{Write stuff}
}
\value{
Path of merged sequences RDS file
}
\description{
Merge multiple sequences as \code{*.rds} in one MSA
}
