% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnabin_functions.R
\name{metrics.lsd}
\alias{metrics.lsd}
\title{Put LSD metrics from list in a table}
\usage{
metrics.lsd(lst = "lsd_results")
}
\arguments{
\item{lst}{List of LSD results}
}
\description{
Put LSD metrics from list in a table
}
