% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{name.list.group}
\alias{name.list.group}
\title{name list of files}
\usage{
name.list.group(lst, regroup = FALSE)
}
\arguments{
\item{lst}{List of paths containing group info}

\item{regroup}{Regroup elements in a list of vectors for each unique group}
}
\value{
Named list
}
\description{
name list of files
}
