% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assortativity.R
\name{newman_r}
\alias{newman_r}
\title{Newman assortativity coefficient from original paper}
\usage{
newman_r(mat)
}
\arguments{
\item{mat}{mixing matrix}
}
\description{
Newman assortativity coefficient from original paper
}
\examples{
##- Table 1
e <- matrix(data = c(0.258,0.016,0.035,0.013,
  0.012,0.157,0.058,0.019,
  0.013,0.023,0.306,0.035,
  0.005,0.007,0.024,0.016),
  nrow = 4, ncol = 4, byrow = TRUE)
rownames(e) <- c("MenBlack", "MenHispanic", "MenWhite", "MenOther")
colnames(e) <- c("WomenBlack", "WomenHispanic", "WomenWhite", "WomenOther")
e
newman_r(e)

##- Bipartite graph
g <- igraph::graph_from_incidence_matrix(e, weighted = TRUE)
igraph::V(g)$color <- igraph::V(g)$type + 1
igraph::plot.igraph(g, edge.width = igraph::E(g)$weight*20)

##- example from Newmann p.12
M <- matrix(data = c(50,50,0,50,50,0,0,0,2)/202, nrow = 3, byrow = TRUE)
M
##- expected value under random allocation:
# $E = \\sum_i{m_{ij}} \\otimes \\sum_ j{m_{ij}} / \\sum_{ij}{m_{ij}}$
E <- rowSums(M) \%o\% colSums(M)/sum(M)
E
##- assortativity matrix
A <- (M - E)/E
A
identical(mat2assortmat(M), A) # TRUE
##- assortativity coef: $r = (\\Tr(M) - \\Tr(E)) / (1 - \\Tr(E))$
r <- (sum(diag(M)) - sum(diag(E)))/ (1 - sum(diag(E)))
r
identical(r, newman_r(M)) # 0.029
}
\references{
Newman, M.E.J., 2003. Mixing patterns in networks. Phys. Rev. E 67, 026126.
\url{https://doi.org/10.1103/PhysRevE.67.026126}
}
