% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{outgroup_choice}
\alias{outgroup_choice}
\title{Selection of subtype reference sequence serving as outgroup}
\usage{
outgroup_choice(subtype)
}
\arguments{
\item{subtype}{subtype of seqs}
}
\description{
Selection of subtype reference sequence serving as outgroup
}
