% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{pipeline_blastdb}
\alias{pipeline_blastdb}
\title{Series of functions to generate a BLAST database}
\usage{
pipeline_blastdb(path.msa = "data/MSM_CRF02AG.rds", keep.splits = TRUE,
  make.blastdb = TRUE, ...)
}
\arguments{
\item{path.msa}{Full path name to RDS or RDA containing DNAbin object}

\item{keep.splits}{Keep the tmp folder containing original and drm-free splits [default = TRUE]}

\item{make.blastdb}{Make BLAST database of final MSA [default = FALSE]}

\item{...}{Options for \code{hpc.cmd.funr}, i.e. turning on/off verbose and choosing HPC queue (\code{hpc.q})}
}
\description{
Script for:
\describe{
  \item{\code{splitForDRM}}{splitting MSA}
  \item{\code{bigphylo_drm}}{drm removal}
  \item{\code{merge_split}}{reassembling MSA}
  \item{\code{make_blast_db}}{make BLAST database from the resulted MSA}
  }
}
