% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_trees.R
\name{pipeline_bstrees}
\alias{pipeline_bstrees}
\title{Pipeline for bootstrap trees}
\usage{
pipeline_bstrees(pathmsa = "data/MLTREE0/MSM_CRF02AG_fltd.fas",
  dirbstree = "data/BSTREES", bs_only = FALSE, BSD = 1, BSE = 2,
  Q = NA)
}
\arguments{
\item{pathmsa}{path to fasta}

\item{dirbstree}{output dir}

\item{bs_only}{turn off LSD steps}

\item{BSD}{debut bootstrap}

\item{BSE}{end bootstrap}

\item{Q}{HPC queue}
}
\description{
Pipeline for bootstrap trees
}
