% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{pipeline_msa}
\alias{pipeline_msa}
\title{Series of functions to generate a MSA ready for maximum likelihood tree}
\usage{
pipeline_msa(path.msa = "data/MSM_CRF02AG.rds", keep.splits = TRUE,
  do.blast = TRUE, do.outgroup = TRUE, pathdb = "blastDB",
  nbysplit = 200, ...)
}
\arguments{
\item{path.msa}{Full path name to RDS or RDA containing DNAbin object}

\item{keep.splits}{Keep the tmp folder containing original and drm-free splits [default = TRUE]}

\item{do.blast}{Blast split sequences, merge unique outputs and align [default = FALSE]}

\item{do.outgroup}{Add outgroup [default = FALSE]}

\item{pathdb}{Path to BLAST database [default = "blastDB"]}

\item{nbysplit}{Size of split for \code{splitForDRM}}

\item{...}{Options for \code{hpc.cmd.funr}, i.e. turning on/off verbose and choosing HPC queue (\code{hpc.q})}
}
\description{
Script for:
\describe{
  \item{\code{splitForDRM}}{1. splitting MSA}
  \item{\code{bigphylo_drm}}{2a. removing drug resistance mutations}
  \item{\code{blast_splits}}{2b. blasting seq splits against corresponding database}
  \item{\code{merge_split}}{3. reassembling MSA}
  \item{\code{merge_blasts}}{4. merging unique BLAST hits}
  \item{\code{prep_and_align_blast}}{5. aligning seq and hits}
  \item{\code{align_mafft}}{6. adding outgroup sequences}
  }
}
