% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_trees.R
\name{pipeline_tree0}
\alias{pipeline_tree0}
\title{Pipeline for tree0 and filtering}
\usage{

  pipeline_tree0(pathmsa = "./data/MSM_CRF02AG_splits/MSM_CRF02AG_nodrm_ref_og.fas",
  dirtree = "data/MLTREE0", Q = NA)
}
\arguments{
\item{pathmsa}{path to fasta}

\item{dirtree}{dir of output}

\item{Q}{HPC queue}
}
\description{
Pipeline for tree0 and filtering
}
