% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnabin_functions.R
\name{plot_all_lsd_tree}
\alias{plot_all_lsd_tree}
\title{from list of LSD results, plot all dated trees at once}
\usage{
plot_all_lsd_tree(lst = "lsd_results_og_nooout")
}
\arguments{
\item{lst}{list of dated trees}
}
\description{
from list of LSD results, plot all dated trees at once
}
