% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnabin_functions.R
\name{plot_tree_col_og}
\alias{plot_tree_col_og}
\title{Plot tree with external sequences colored}
\usage{
plot_tree_col_og(tr, tmrca = NULL, title = NULL, tiplab = FALSE)
}
\arguments{
\item{tr}{phylo tree}

\item{tmrca}{date of root to add time axis}

\item{title}{title of plot}

\item{tiplab}{logical to show tip labels}
}
\description{
Plot tree with external sequences colored
}
