% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_w_reg}
\alias{plot_w_reg}
\title{XY bubble plot of weighted regression}
\usage{
plot_w_reg(df, x = "donor_v", y = "recip_v", w = "ip", title = "",
  labx = "", laby = "", labz = "", lims = NULL, alpha = TRUE,
  thr_w = NULL, qt = 0.975, regline = TRUE, FUN = function(x)
  high_contrast(sum(x)), nbins = 30, collo = "white", colhi = "black",
  diverging = FALSE, type = c("bubble", "hexplot"), verbose = TRUE)
}
\arguments{
\item{df}{dataframe with x,y, weights}

\item{x}{x}

\item{y}{y}

\item{w}{weights}

\item{title}{title passed to ggplot}

\item{labx}{name x}

\item{laby}{name y}

\item{labz}{name of legend}

\item{lims}{optional imposed limits for x & y axis}

\item{alpha}{logical, control transparency}

\item{thr_w}{Threshold of weights above which values are plotted. If \code{NULL}, plot weights above \code{qt} percentile. Set to \code{-Inf} to plot everything}

\item{qt}{probability for percentile threshold [default=0.975]}

\item{regline}{logical, calculate and plot regression line}

\item{FUN}{function for \code{\link[ggplot2]{stat_summary_hex}}}

\item{nbins}{for hexplot}

\item{collo}{colour low level}

\item{colhi}{colour low level}

\item{diverging}{diverging colour gradient with center 0}

\item{type}{bubble (plot) or hexplot}

\item{verbose}{details}
}
\description{
XY bubble plot of weighted regression
}
