% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assortativity.R
\name{plots_grid}
\alias{plots_grid}
\title{Don't know what this does}
\usage{
plots_grid(mat, contrast = 1, transpose = TRUE, ...)
}
\arguments{
\item{mat}{matrix}

\item{contrast}{change heatmap contrast (independent of plotted values)}

\item{transpose}{transpose matrix}

\item{...}{extra parms for lattice::levelplot}
}
\description{
yet
}
