% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnabin_functions.R
\name{prep_and_align_blast}
\alias{prep_and_align_blast}
\title{Process MSA and align}
\usage{

  prep_and_align_blast(infile = "data/MSM_CRF02AG_splits/MSM_CRF02AG_nodrm.rds")
}
\arguments{
\item{infile}{Path to MSA}
}
\value{
Path to output
}
\description{
combine \code{prep_blast_align} and \code{align_mafft}
}
