% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnabin_functions.R
\name{rds2fasta}
\alias{rds2fasta}
\title{Convert DNAbin in fasta}
\usage{
rds2fasta(infile = "test.rds", verbose = TRUE)
}
\arguments{
\item{infile}{Path to RDS file containing DNAbin object}

\item{verbose}{Indicates path of output file}
}
\value{
Path to fasta file
}
\description{
Convert in place \code{*.rds} to \code{*.fas}
}
