% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnabin_functions.R
\name{redo_lsd}
\alias{redo_lsd}
\title{Filter LSD tree for tips producing MRCA ancestral to global sequences and re date new tree}
\usage{
redo_lsd(pathlsdtree, outgroup = TRUE, ...)
}
\arguments{
\item{pathlsdtree}{Path to LSD result
e.g. './BS_trees/MSM_CRF02AG/
ExaML_result.MSM_CRF02AG.finaltree.001_lsd.rds'}

\item{outgroup}{Outgroup option of LSD (LSD outputs a tree without outgroup)}

\item{...}{Passed to hpc.cmd.funr (i.e. hpc.q = 'pqeelab', verbose = FALSE)}
}
\description{
Filter LSD tree for tips producing MRCA ancestral to global sequences and re date new tree
}
