% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnabin_functions.R
\name{rm_mrca_outliers}
\alias{rm_mrca_outliers}
\title{Remove mrca outliers from tree}
\usage{
rm_mrca_outliers(TREE, outgroup = TRUE, OUTFILE = NULL, verbose = TRUE)
}
\arguments{
\item{TREE}{Path to tree}

\item{outgroup}{Outgroup option of LSD (LSD outputs a tree without outgroup)}

\item{OUTFILE}{Path to save list of outliers, in rds}

\item{verbose}{Details}
}
\value{
Path to filtered tree
}
\description{
Remove mrca outliers from tree
}
