% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnabin_functions.R
\name{rm_outliers_msa}
\alias{rm_outliers_msa}
\title{Remove outliers from MSA in DNAbin}
\usage{
rm_outliers_msa(pathmsa, pathoutliers, suffix = "fltd", verbose = TRUE)
}
\arguments{
\item{pathmsa}{Path to MSA saved as \code{seq} DNAbin object}

\item{pathoutliers}{Path to outliers contained in labels of \code{seq}, vector saved in RDS format}

\item{suffix}{Suffix added to filename}

\item{verbose}{print details}
}
\value{
save a fasta file
}
\description{
Remove outliers from MSA in DNAbin
}
