% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnabin_functions.R
\name{rm_rtt_outliers}
\alias{rm_rtt_outliers}
\title{Remove RTT outliers from tree}
\usage{
rm_rtt_outliers(TREE, outgroup = TRUE, OUTFILE = NULL, rooting = TRUE,
  verbose = TRUE)
}
\arguments{
\item{TREE}{Path to tree}

\item{outgroup}{Outgroup option of LSD (LSD outputs a tree without outgroup)}

\item{OUTFILE}{Path to save list of outliers, in rds}

\item{rooting}{Passed to dates_of_rtree}

\item{verbose}{Details}
}
\value{
Path to filtered tree
}
\description{
Remove RTT outliers from tree
}
