% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assortativity.R
\name{rorder}
\alias{rorder}
\title{reorder matrix factor level}
\usage{
rorder(m, o, normalize = FALSE)
}
\arguments{
\item{m}{matrix}

\item{o}{character vector of rownames/colnames}

\item{normalize}{logical to normalize if matrix subsetted}
}
\description{
reorder matrix factor level
}
