% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_functions.R
\name{sa_by_clades}
\alias{sa_by_clades}
\title{Compute infector probabilities on clades}
\usage{
sa_by_clades(tr, parms_SA = get_sa_parms_range(), p = 1)
}
\arguments{
\item{tr}{phylo tree}

\item{parms_SA}{Parameters from \code{\link{get_sa_parms_range}}}

\item{p}{proportion of subtype represented (useless)}
}
\value{
dataframe(donor, recip, ip)
}
\description{
Compute infector probabilities on clades
}
\details{
\code{phydynR::phylo.source.attribution.hiv.msm} wraps \code{phydynR::phylo.source.attribution.multiDeme.fgy} that runs \code{phydynR::sourceAttribMultiDemeCpp} (mode = 1) or \code{phydynR::sourceAttribMultiDemeCpp2} (mode = 2)
}
