% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{select_cohort}
\alias{select_cohort}
\title{Make subset MSA for samples collected as of year x}
\usage{
select_cohort(pathmsa, year = 2010)
}
\arguments{
\item{pathmsa}{path to RDS file of MSA}

\item{year}{first year of selection}
}
\description{
From year contained in \code{dateres} in \code{df}
}
