% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_mld.R
\name{single_mc_mld}
\alias{single_mc_mld}
\title{Monte-Carlo simulation of likely donor to each recipient}
\usage{
single_mc_mld(x = W)
}
\arguments{
\item{x}{Edge list dataframe of infector probabilities with columns (donor, recip, ip)}
}
\value{
Vector of MLD count named by donor ID
}
\description{
Counts how many times each donor is the most likely donor to a recipient.
For each recipient i, one donor or "outside source" is sampled with probabilities = infector probs for each link  + (1 - indegree)
}
\examples{
d <- as.character(1:3)
r <- as.character(1:3)
W <- expand.grid('donor' = d, 'recip' = r, stringsAsFactors = FALSE)
W$ip <- runif(dim(W)[1])/length(d) # arbitrary ip
W # edge list
tapply(W$ip, W$recip, sum) # in-degree for each recipient
Nsim=10
replicate(Nsim, single_mc_mld(x = W))
}
