% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnabin_functions.R
\name{splitForDRM}
\alias{splitForDRM}
\title{Split MSA in subsets}
\usage{
splitForDRM(path.seqs, path.out = ".", path.hxb2 = system.file("ext",
  "hxb2.txt", package = "garel", mustWork = TRUE), n.by.split = 100,
  verbose = TRUE)
}
\arguments{
\item{path.seqs}{Path to RDA or RDS file containing large alignment in DNAbin format}

\item{path.out}{Path of folder (based on \code{path.seqs}) to save the splitted MSAs}

\item{path.hxb2}{Path to HXB2 txt file (default = location in \code{garel} package)}

\item{n.by.split}{Number of sequences by split (default = 100)}

\item{verbose}{blabla}
}
\value{
List of splits and save splits as RDS in \code{path.out}
}
\description{
Split MSA in subsets, each aligned with HXB2
}
