% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_msas.R
\name{stratify_msa}
\alias{stratify_msa}
\title{Extract MSAs stratified by main group/subtype}
\usage{
stratify_msa()
}
\value{
Save MSAs as RDS in \code{data}
}
\description{
Extract MSAs stratified by main group/subtype
}
